feat(server): split review and social endpoints
